requirements.txt
numpy
torch
torchvision
pillow
scipy
maptlotlib
pandas
